sources/producer/requirements.txt
kafka-python==2.0.1
twitter==1.18.0